Clear the list first if items don't have unique IDs
Replacing the entire list at the end of fetching new items helps reduce
flicker, but if the items don't have unique-IDs, the UI will not update
correctly. So if the list items don't have unique IDs, we must clear the
list before we fetch new items. This will cause the UI to flicker, but
it will update correctly.