Rewrite network-graph component to use Vue reactivity
Previously, the network-graph Vue component used D3JS to layout the
graph, but used Vue components to layout the contents of the nodes
in the graph. The network-graph component had to take care of the
Vue lifecycle for the node components. It became harder to maintain
this as the feature-set grew.

So the network-graph component has been almost completely rewritten
to rely solely on Vue and its reactivity throughout. D3JS is now
limited to computing the SVG paths for the network links.