Render links the first time the graph is mounted
For links to get rendered in the correct positions, nodes need to be
rendered so that we can get their DOM positions and position the links
accordingly. This was not happeing because link positioning requires a
reference to the div containing the nodes. This reference is not
available initially, resulting in links not getting rendered initially.
This commit passes the initial refernce to methods in child and grand-
child components.