Update a group's children-list based on pinning
If a group is expanded and all its children are visible as separate nodes,
then pinning can result in some of those children being hidden.
So when a group is collapsed, it should not list all its children; it
should only list the children that would have been visible if the group
had been expanded. That is what this commit does.