Expand links svg as nodes div expands
If node columns are added after initial graph rendering, the links svg
needs to expand accordingly. For this to happen, the parent div of the
nodes and links needs to expand. This happens only if the parent is an
inline-block, but the parent must be flex for the links to be overlaid
on top of the nodes. So an inline-block wrapper div is needed for all
of it work.